Add new examples & better export

+++ Assets/Editor/BackgroundEditor.cs
@@ -1,6 +1,8 @@
-﻿using System.IO;
+﻿using System.Collections.Generic;
+using System.IO;
 using UnityEditor;
 using UnityEngine;
+using UnityEngine.Video;
 using TrombLoader.Data;
 
 namespace TrombLoader 
@@ -45,21 +47,64 @@ namespace TrombLoader
                         string fileName = Path.GetFileName(path);
                         string folderPath = Path.GetDirectoryName(path);
 
+                        int serializedCount = 0;
+                        // serialize video clips because unity REALLY does not like making them work in assetbundles
+                        foreach (var videoPlayer in clonedTromboneBackground.gameObject.GetComponentsInChildren<VideoPlayer>())
+                        {
+                            if (videoPlayer.clip != null)
+                            {
+                                // handle VideoClip.originalClip returning an invalid file extension sometimes
+                                foreach (var file in Directory.GetFiles(Path.GetDirectoryName(videoPlayer.clip.originalPath)))
+                                {
+                                    if (videoPlayer.clip == null) break;
+
+                                    if (!file.EndsWith(".meta") && Path.GetFileNameWithoutExtension(file) == Path.GetFileNameWithoutExtension(videoPlayer.clip.originalPath))
+                                    {
+                                        videoPlayer.clip = null;
+                                        videoPlayer.url = $"SERIALIZED_OUTSIDE_BUNDLE/SERIALIZED_{serializedCount}_" + Path.GetFileName(file);
+
+                                        var newVideoPath = Path.Combine(folderPath, $"SERIALIZED_{serializedCount}_" + Path.GetFileName(file));
+                                        
+                                        if (File.Exists(newVideoPath)) File.Delete(newVideoPath);
+                                        File.Copy(file, newVideoPath);
+
+                                        serializedCount++;
+                                        break;
+                                    }
+                                }
+                            }
+                        }
+
+                        // serialize tromboners (this one is not unity's fault, it's base game weirdness)
+                        List<string> trombonePaths = new List<string>(){"Assets/_Background.prefab"};
+                        
+                        int instanceID = 0;
+                        
+                        foreach (var tromboner in clonedTromboneBackground.gameObject.GetComponentsInChildren<TrombonerPlaceholder>())
+                        {
+                            tromboner.InstanceID = instanceID;
+                            tromboner.gameObject.name = $"_{instanceID}";
+                            
+                            instanceID++;
+                        }
+                        
                         PrefabUtility.SaveAsPrefabAsset(clonedTromboneBackground.gameObject, "Assets/_Background.prefab");
                         AssetBundleBuild assetBundleBuild = default;
                         assetBundleBuild.assetBundleName = fileName;
-                        assetBundleBuild.assetNames = new string[] {"Assets/_Background.prefab"};
+                        assetBundleBuild.assetNames = trombonePaths.ToArray();
 
                         BuildPipeline.BuildAssetBundles(Application.temporaryCachePath,
                             new AssetBundleBuild[] {assetBundleBuild}, BuildAssetBundleOptions.ForceRebuildAssetBundle,
                             EditorUserBuildSettings.activeBuildTarget);
                         EditorPrefs.SetString("currentBuildingAssetBundlePath", folderPath);
                         EditorUserBuildSettings.SwitchActiveBuildTarget(selectedBuildTargetGroup, activeBuildTarget);
+                        
+                        foreach (var asset in trombonePaths)
+                        {
+                            AssetDatabase.DeleteAsset(asset);
+                        }
 
-                        AssetDatabase.DeleteAsset("Assets/_Background.prefab");
-
-                        if (File.Exists(path))
-                            File.Delete(path);
+                        if (File.Exists(path)) File.Delete(path);
 
                         // Unity seems to save the file in lower case, which is a problem on Linux, as file systems are case sensitive there
                         File.Move(Path.Combine(Application.temporaryCachePath, fileName.ToLowerInvariant()), path);
